# TriPrize/.specify/001-triangle-lottery/contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: Triangle Lottery Sales API
  description: |
    REST API for triangle-shaped lottery campaign platform.

    ## Authentication
    All authenticated endpoints require Firebase ID Token in Authorization header:
    ```
    Authorization: Bearer <firebase_id_token>
    ```

    ## Idempotency
    POST/PUT/DELETE requests support idempotency via `Idempotency-Key` header:
    ```
    Idempotency-Key: <client-generated-uuid>
    ```

    ## Error Handling
    All error responses follow this format:
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message",
        "details": {}
      }
    }
    ```
  version: 1.0.0
  contact:
    name: TriPrize API Support
    email: support@triprize.example.com
servers:
  - url: https://api.triprize.example.com/v1
    description: Production server
  - url: https://api-staging.triprize.example.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - FirebaseAuth: []
paths:
...
  /webhooks/stripe:
    post:
      tags: [Webhooks]
      summary: Stripe webhook handler
      description: Handle Stripe webhook events (FR-022)
      security: []
      parameters:
        - name: Stripe-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed successfully
        '400':
          description: Invalid signature or payload
        '500':
          description: Processing error
components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID Token
...